Update code to fix bug with BT connection.
There is a bug in the Nordic UART protocol implementation that causes a
memory error when there's a disconnect (I think it's when the disconnect
occurs during uart.write). This adds a 5 second watchdog that resets the device
when the error happens. It also adds a safemode.py file to reset device
when safemode is entered for reasons other than the user pressing the
reset button (since the memory error causes safemode).